DHFPROD-5155: Refactor test data/configurations and mock responses for UI unit tests
