+++ marklogic-data-hub-central/ui/src/assets/mock-data/advanced-settings.data.ts
