+++ marklogic-data-hub-central/ui/src/assets/mock-data/common.data.ts
