+++ marklogic-data-hub-central/ui/src/assets/mock-data/flows.data.ts
