+++ marklogic-data-hub-central/ui/src/assets/mock-data/ingestion.data.ts
