+++ marklogic-data-hub-central/ui/src/components/entities/entity-tiles.test.tsx
@@ -4,7 +4,7 @@ import {fireEvent, render} from '@testing-library/react';
 
 import EntityTiles from './entity-tiles';
 import axiosMock from 'axios'
-import data from "../../config/run.config";
+import data from "../../assets/mock-data/flows.data";
 import {act} from "react-dom/test-utils";
 
 jest.mock('axios');
