+++ marklogic-data-hub-central/ui/src/components/entities/mapping/create-edit-mapping-dialog/create-edit-mapping-dialog.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, fireEvent, cleanup, wait } from '@testing-library/react';
 import CreateEditMappingDialog from './create-edit-mapping-dialog';
-import data from "../../../../config/test-data.config";
+import data from "../../../../assets/mock-data/common.data";
 
 describe('Create/Edit Mapping artifact component', () => {
