+++ marklogic-data-hub-central/ui/src/components/tiles/tiles.test.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 import Tiles from './tiles';
 import { faCube } from "@fortawesome/free-solid-svg-icons";
-import TestComponent from '../../config/test-component'
+import TestComponent from '../../assets/mock-data/test-component'
 
 describe('Tiles component', () => {
     test('renders with a FontAwesome icon', () => {
